feat: 마이페이지 홈 -> 상품 주문 내역 page html 수정

+++ src/main/java/com/bitcamp/sc/domain/mypage/domain/OrderList.java
@@ -8,11 +8,11 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class OrderList {
-	private int pidx;
+	private long pidx;
 	private int pprice;
 	private String pdate;
 	private String pstatus;
-	private int oidx;
+	private long oidx;
 	private int amount;
 	private String gtitle;
 	private String gname;
